requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap4==2024.1
Django==5.1.1
django-crispy-forms==2.3
django-environ==0.11.2
django-mailgun-mime==0.1.7
idna==3.10
pillow==10.4.0
py4j==0.10.9.7
pyspark==3.5.1
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.11.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
urllib3==2.2.3
